Extract render code from PdfHelper into Renderer class
